+++ src/connect.js
@@ -1,4 +1,5 @@
 import React, { Component, PropTypes } from 'react';
+import invariant from 'invariant';
 import ApolloPassport from 'apollo-passport/lib/client';
 
 function getDisplayName(WrappedComponent) {
@@ -15,10 +16,36 @@ export default function apolloPassportConnect(mapStateToProps) {
     class ApolloPassportConnect extends Component {
       constructor(props, context) {
         super(props, context);
-        const apolloPassport = this.apolloPassport = context.apolloPassport;
+        const apolloPassport = this.apolloPassport
+          = context.apolloPassport || props.apolloPassport;
 
-        this.state = { auth: apolloPassport.getState() };
-        apolloPassport.subscribe(auth => this.setState({ auth }));
+        /*
+        invariant(apolloPassport,
+          `Could not find "apolloPassport" in either the context or ` +
+          `props of "${connectDisplayName}". ` +
+          `Either wrap the root component in an <ApolloPassportProvider>, ` +
+          `or explicitly pass "apolloPassport" as a prop to "${connectDisplayName}".`
+        )
+        */
+
+        if (!apolloPassport) {
+          // TODO this can happen with SSR, need to think about this more
+          // i.e. can we show the login state?
+          this.state = {
+            auth: {
+              data: {}
+            }
+          };
+
+          this.actions = {};
+          this.apolloPassport = {};
+
+          return;
+        }
+
+
+        this.state = { auth: apolloPassport.getState(), apolloPassport };
+        apolloPassport.subscribe(auth => this.setState({ auth, apolloPassport }));
       }
 
       componentWillUnmount() {
@@ -37,12 +64,15 @@ export default function apolloPassportConnect(mapStateToProps) {
     ApolloPassportConnect.displayName = connectDisplayName;
     ApolloPassportConnect.WrappedComponent = ComponentToWrap;
 
+    ApolloPassportConnect.propTypes = {
+      apolloPassport: PropTypes.instanceOf(ApolloPassport)
+    };
+
     ApolloPassportConnect.contextTypes = {
       apolloPassport: PropTypes.instanceOf(ApolloPassport)
-    }
+    };
 
     return ApolloPassportConnect;
-
   }
 
 }